--HG--
branch : supervisor-read-layer


Former-commit-id: 09639f396a6d8806de605afb11bc46a9b28a5cb5

+++ RavenQuestionnaire.Web/Main.Core/Entities/SubEntities/Complete/Question/NumericCompleteQuestion.cs
@@ -15,7 +15,7 @@ namespace Main.Core.Entities.SubEntities.Complete.Question
     /// <summary>
     /// The numeric complete question.
     /// </summary>
-    public sealed class NumericCompleteQuestion : AbstractCompleteQuestion, INumericQuestion, ICompelteValueQuestion<int?>
+    public sealed class NumericCompleteQuestion : AbstractCompleteQuestion, INumericQuestion, ICompelteValueQuestion<double?>
     {
 
         #region Constructors and Destructors
@@ -138,8 +138,8 @@ namespace Main.Core.Entities.SubEntities.Complete.Question
             }
             else
             {
-                int value;
-                if (int.TryParse(answerValue.Trim(), out value))
+                double value;
+                if (double.TryParse(answerValue.Trim(), out value))
                 {
                     this.Answer = value;
                 }
@@ -150,7 +150,7 @@ namespace Main.Core.Entities.SubEntities.Complete.Question
 
         #region Implementation of ICompelteValueQuestion<int>
 
-        public int? Answer { get; set; }
+        public double? Answer { get; set; }
 
         #endregion
     }
